refactor to pass document id
