--- src/packages/core/modal/token/allowed-document-types-modal.token.ts
@@ -1,18 +1,0 @@
-import { UmbModalToken } from '@umbraco-cms/backoffice/modal';
-
-export interface UmbAllowedDocumentTypesModalData {
-	parentId: string | null;
-	parentName?: string;
-}
-
-export interface UmbAllowedDocumentTypesModalResult {
-	documentTypeKey: string;
-}
-
-export const UMB_ALLOWED_DOCUMENT_TYPES_MODAL = new UmbModalToken<
-	UmbAllowedDocumentTypesModalData,
-	UmbAllowedDocumentTypesModalResult
->('Umb.Modal.AllowedDocumentTypes', {
-	type: 'sidebar',
-	size: 'small',
-});
