+++ src/packages/core/modal/token/create-document-modal.token.ts
@@ -1,0 +1,17 @@
+import { UmbModalToken } from '@umbraco-cms/backoffice/modal';
+
+export interface UmbCreateDocumentModalData {
+	id: string | null;
+}
+
+export interface UmbCreateDocumentModalResult {
+	documentTypeId: string;
+}
+
+export const UMB_CREATE_DOCUMENT_MODAL = new UmbModalToken<UmbCreateDocumentModalData, UmbCreateDocumentModalResult>(
+	'Umb.Modal.CreateDocument',
+	{
+		type: 'sidebar',
+		size: 'small',
+	},
+);
