+++ src/packages/core/modal/token/index.ts
@@ -1,5 +1,5 @@
 export * from './modal-token.js';
-export * from './allowed-document-types-modal.token.js';
+export * from './create-document-modal.token.js';
 export * from './change-password-modal.token.js';
 export * from './code-editor-modal.token.js';
 export * from './confirm-modal.token.js';
